页面修改

+++ src/main/resources/static/js/community.js
@@ -3,9 +3,9 @@
  * 提交回复
  */
 function post() {
-    var questionId = $("#question_id").val();
+    var articleId = $("#article_id").val();
     var content = $("#comment_content").val();
-    comment2target(questionId, 1, content);
+    comment2target(articleId, 1, content);
 }
 
 function comment2target(targetId, type, content) {
@@ -30,7 +30,41 @@ function comment2target(targetId, type, content) {
                 if (response.code == 2003) {
                     var isAccepted = confirm(response.message);
                     if (isAccepted) {
-                        window.open("https://github.com/login/oauth/authorize?client_id=54920a01aa826eaee59f&redirect_uri=" + document.location.origin + "/callback&scope=user&state=1");
+                        // window.open("https://github.com/login/oauth/authorize?client_id=54920a01aa826eaee59f&redirect_uri=" + document.location.origin + "/callback&scope=user&state=1");
+                        window.open("/login");
+                        window.localStorage.setItem("closable", true);
+                    }
+                } else {
+                    alert(response.message);
+                }
+            }
+        },
+        dataType: "json"
+    });
+}
+
+function addLike() {
+    var articleId = $("#article_id").val();
+    var creatorId=$("#article_creator").val();
+addL(articleId,creatorId);
+}
+function addL(articleId,creatorId) {
+    $.ajax({
+        type: "POST",
+        url: "/like",
+        contentType: 'application/json',
+        data: JSON.stringify({
+            "id":articleId,
+            "creatorId":creatorId
+        }),
+        success: function (response) {
+            if (response.code == 200) {
+                window.location.reload();
+            } else {
+                if (response.code == 2003) {
+                    var isAccepted = confirm(response.message);
+                    if (isAccepted) {
+                        window.open("/login");
                         window.localStorage.setItem("closable", true);
                     }
                 } else {
@@ -75,7 +109,8 @@ function comment3target(targetId, type, content,commentParentId) {
                 if (response.code == 2003) {
                     var isAccepted = confirm(response.message);
                     if (isAccepted) {
-                        window.open("https://github.com/login/oauth/authorize?client_id=54920a01aa826eaee59f&redirect_uri=" + document.location.origin + "/callback&scope=user&state=1");
+                        // window.open("https://github.com/login/oauth/authorize?client_id=54920a01aa826eaee59f&redirect_uri=" + document.location.origin + "/callback&scope=user&state=1");
+                        window.open("/login");
                         window.localStorage.setItem("closable", true);
                     }
                 } else {
